+++ horned-creatures-lab/src/imageItem.js
@@ -1,4 +1,4 @@
-import React from 'react'
+import React from 'react';
 
 export default class ImageItem extends React.Component {
     render() {
